Using faiss gpu for k-means, add task split results for infoseek

+++ data_utils.py
@@ -4,12 +4,10 @@ import numpy as np
 from sklearn.cluster import KMeans
 import asyncio
 import pandas as pd
+import faiss
 
 from constant import DATA_DIR
 from embedder import OpenAIEmbedder
-
-
-
 
 async def split_infoseek(n_clusters=0):
     
@@ -22,18 +20,22 @@ async def split_infoseek(n_clusters=0):
         questions = df['question'].tolist()
 
         # Generate embeddings using the embedder
-        print(f"Processing split: {split}, {len(questions)} questions")
-        embeddings = await embedder(questions)
+        if os.path.exists(embedding_path):
+            print(f"Loading embedding: {split}, {len(questions)} questions")
+            embeddings = np.load(embedding_path)
+        else:
+            print(f"Processing split: {split}, {len(questions)} questions")
+            embeddings = await embedder(questions)
 
-        # Save embeddings to a .npy file
-        np.save(embedding_path, embeddings)
+            # Save embeddings to a .npy file
+            np.save(embedding_path, embeddings)
 
         # Add embeddings to the dataframe for clustering
         df['embedding'] = list(embeddings)
         df['split'] = split
         return df
 
-    embedder = OpenAIEmbedder(use_batch_api=False)
+    embedder = OpenAIEmbedder(use_batch_api=True)
     tasks = []
 
     for split in ('train', 'val'):
@@ -47,14 +49,18 @@ async def split_infoseek(n_clusters=0):
 
     # Number of clusters
     if n_clusters == 0:
-        n_clusters = len(combined_df) // 100
+        n_clusters = max(len(combined_df) // 100, 2)
 
-    # Perform clustering
-    print("Performing kmeans...")
-    matrix = np.vstack(combined_df['embedding'].values)
-    kmeans = KMeans(n_clusters=n_clusters, init='k-means++', random_state=42)
-    kmeans.fit(matrix)
-    combined_df['Cluster'] = kmeans.labels_
+    # Perform clustering using Faiss
+    print("Performing Faiss KMeans clustering...")
+    matrix = np.vstack(combined_df['embedding'].values).astype('float32')
+    res = faiss.StandardGpuResources()
+    kmeans = faiss.Kmeans(d=matrix.shape[1], k=n_clusters, niter=20, verbose=True, gpu=True)
+    kmeans.train(matrix)
+    
+    # Assign clusters to data points
+    _, labels = kmeans.index.search(matrix, 1)
+    combined_df['Cluster'] = labels.flatten()
 
     # Drop the embeddings column
     combined_df = combined_df.drop(columns=['embedding'])
@@ -77,13 +83,17 @@ async def split_okvqa(n_clusters=0):
         del data
 
         questions = df['question'].tolist()
-        print(f"Processing OKVQA split: {split}, {len(questions)} questions")
 
         # Generate embeddings using the embedder
-        embeddings = await embedder(questions)
+        if os.path.exists(embedding_path):
+            print(f"Loading embedding: {split}, {len(questions)} questions")
+            embeddings = np.load(embedding_path)
+        else:
+            print(f"Processing OKVQA split: {split}, {len(questions)} questions")
+            embeddings = await embedder(questions)
 
-        # Save embeddings to a .npy file
-        np.save(embedding_path, embeddings)
+            # Save embeddings to a .npy file
+            np.save(embedding_path, embeddings)
 
         # Add embeddings to the dataframe for clustering
         df['embedding'] = list(embeddings)
@@ -104,21 +114,25 @@ async def split_okvqa(n_clusters=0):
 
     # Number of clusters
     if n_clusters == 0:
-        n_clusters = len(combined_df) // 100
-
-    # Perform clustering
-    print("Performing kmeans...")
-    matrix = np.vstack(combined_df['embedding'].values)
-    kmeans = KMeans(n_clusters=n_clusters, init='k-means++', random_state=42)
-    kmeans.fit(matrix)
-    combined_df['Cluster'] = kmeans.labels_
+        n_clusters = max(len(combined_df) // 100, 2)
+    
+    # Perform clustering using Faiss
+    print("Performing Faiss KMeans clustering...")
+    matrix = np.vstack(combined_df['embedding'].values).astype('float32')
+    res = faiss.StandardGpuResources()
+    kmeans = faiss.Kmeans(d=matrix.shape[1], k=n_clusters, niter=20, verbose=True, gpu=True)
+    kmeans.train(matrix)
+    
+    # Assign clusters to data points
+    _, labels = kmeans.index.search(matrix, 1)
+    combined_df['Cluster'] = labels.flatten()
 
     # Drop the embeddings column
     combined_df = combined_df.drop(columns=['embedding'])
 
     # Save the clustered dataframe to a JSONL file
     print("Saving results...")
-    save_path = os.path.join('data', f"OpenEnded_mscoco_trainval2014_questions.jsonl")
+    save_path = os.path.join('data', f"OpenEnded_mscoco_trainval2014_questions.clustered.jsonl")
     combined_df.to_json(save_path, orient='records', lines=True)
